feat(locale): use translated descriptions for launch argument search
