+++ src/routes/Servers/components/Forms/LaunchArguments/LaunchArgumentsForm.tsx
@@ -12,6 +12,7 @@ import { editServerLoadout } from '@/api'
 import { toast } from 'sonner'
 import { defaultLaunchArguments } from './setup/LaunchArguments'
 import { LaunchArgumentFormBuilder } from './LaunchArgumentFormBuilder/LaunchArgumentFormBuilder'
+import { LaunchArgumentsSearch } from './LaunchArgumentsSearch'
 
 const formSchema = z.object({
   common: z
@@ -95,42 +96,10 @@ export function LaunchArgumentForm({
     }
   }
 
-  function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
-    const info = Object.keys(defaultLaunchArguments).map((x) => {
-      // @ts-ignore
-      const filtered_fields = Object.keys(defaultLaunchArguments[x])
-        .filter((key) => key.toLowerCase().includes(e.target.value))
-        .reduce((obj, key) => {
-          // @ts-ignore
-          obj[key] = defaultLaunchArguments[x][key]
-          return obj
-        }, {})
-
-      return {
-        name: x,
-        values: filtered_fields,
-      }
-    })
-
-    const combinedObject = {}
-    info.forEach((x) => {
-      // @ts-ignore
-      combinedObject[x.name] = x.values
-    })
-
-    setFilteredArgs(() => combinedObject)
-  }
-
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(onSubmit)} className="m-auto max-w-screen-md">
-        <input
-          type="text"
-          placeholder={`Search Launch Arguments     [CTRL + F]`}
-          className="fixed top-0 w-[720px] rounded-md border border-gray-500 bg-black p-2 focus:border-cyan-300 focus:outline-none focus:ring-0"
-          onChange={handleChange}
-          ref={searchRef}
-        />
+        <LaunchArgumentsSearch searchRef={searchRef} setFilteredArgs={setFilteredArgs} />
 
         <div className="flex flex-col gap-6 pt-12">
           <LaunchArgumentFormBuilder
